requirements.txt
beautifulsoup4==4.10.0
faiss_cpu==1.7.2
fast_autocomplete==0.9.0
Flask==2.0.3
Flask_RESTful==0.3.9
nltk==3.6.7
pandas==1.3.5
pymongo==4.0.1
rapidfuzz==2.0.7
requests==2.26.0
sentence_transformers==2.2.0
